Refactor: move get-config-for-method to a file
